Added new file format and a handler
In detail, the Main.yml became Main.json. Why? In order to avoid that if this whole story is published here, no extra file converter or the like is needed. A handler has also been added, the purpose of which is to execute functions or the same step by step. What purpose does this serve? Quite simply, it builds on the function of including the script in the header, but independently, so that it could also be used for other things here in the program.